[fix] add faust and kafka as optional extras
