+++ requirements.txt
@@ -1,7 +1,5 @@
 celery>4,<5
 flask<2,>=0.12.3
 python-dateutil>=2.6.1,<3
-kafka-python<2,>=1
 statsd>=3.2.1,<4
 marshmallow>=2.15,<4
-faust>1,<2
